+++ core-services/egov-hrms/src/main/java/org/egov/hrms/service/EmployeeService.java
@@ -106,6 +106,9 @@ public class EmployeeService {
 	@Autowired
 	private IndividualService individualService;
 
+	@Autowired
+	private IndividualService individualService;
+
 	/**
 	 * Service method for create employee. Does following:
 	 * 1. Sets ids to all the objects using idgen service.
@@ -118,15 +121,20 @@ public class EmployeeService {
 	 */
 	public EmployeeResponse create(EmployeeRequest employeeRequest) {
 		RequestInfo requestInfo = employeeRequest.getRequestInfo();
+		String tenantId = employeeRequest.getEmployees().stream().findAny().get().getTenantId();
 		Map<String, String> pwdMap = new HashMap<>();
 		idGenService.setIds(employeeRequest);
 		employeeRequest.getEmployees().stream().forEach(employee -> {
-			enrichCreateRequest(employee, requestInfo);
+			enrichCreateRequest(tenantId, employee, requestInfo);
 			createUser(employee, requestInfo);
 			pwdMap.put(employee.getUuid(), employee.getUser().getPassword());
-			employee.getUser().setPassword(null);
 		});
-		hrmsProducer.push(propertiesManager.getSaveEmployeeTopic(), employeeRequest);
+		hrmsProducer.push(propertiesManager.getHrmsEmailNotifTopic(), employeeRequest);
+
+		// Setting password as null after sending employeeRequest to email notification topic to send email.
+		employeeRequest.getEmployees().forEach(employee -> employee.getUser().setPassword(null));
+
+		hrmsProducer.push(tenantId, propertiesManager.getSaveEmployeeTopic(), employeeRequest);
 		notificationService.sendNotification(employeeRequest, pwdMap);
 		return generateResponse(employeeRequest);
 	}
@@ -140,6 +148,7 @@ public class EmployeeService {
 	 */
 	public EmployeeResponse search(EmployeeSearchCriteria criteria, RequestInfo requestInfo) {
 		boolean  userChecked = false;
+		Long totalCount = 0L;
 		/*if(null == criteria.getIsActive() || criteria.getIsActive())
 			criteria.setIsActive(true);
 		else
@@ -159,6 +168,7 @@ public class EmployeeService {
 				userSearchCriteria.put(HRMSConstants.HRMS_USER_SEARCH_CRITERA_USERNAME, criteria.getCodes().get(0));
 			}
             UserResponse userResponse = userService.getUser(requestInfo, userSearchCriteria);
+			totalCount = userResponse.getTotalCount();
 			userChecked =true;
             if(!CollectionUtils.isEmpty(userResponse.getUser())) {
                  mapOfUsers.putAll(userResponse.getUser().stream()
@@ -181,6 +191,7 @@ public class EmployeeService {
 					userSearchCriteria.put(HRMSConstants.HRMS_USER_SEARCH_CRITERA_TENANTID,criteria.getTenantId());
 					userSearchCriteria.put(HRMSConstants.HRMS_USER_SEARCH_CRITERA_NAME,name);
 					UserResponse userResponse = userService.getUser(requestInfo, userSearchCriteria);
+					totalCount = userResponse.getTotalCount();
 					userChecked =true;
 					if(!CollectionUtils.isEmpty(userResponse.getUser())) {
 						mapOfUsers.putAll(userResponse.getUser().stream()
@@ -194,15 +205,41 @@ public class EmployeeService {
 				else
 					criteria.setUuids(userUUIDs);
 			}
+
+			if(!CollectionUtils.isEmpty(criteria.getUserServiceUuids())) {
+				List<String> userUUIDs = new ArrayList<>();
+				Map<String, Object> userSearchCriteria = new HashMap<>();
+
+				userSearchCriteria.put(HRMSConstants.HRMS_USER_SERACH_CRITERIA_USERTYPE_CODE, HRMSConstants.HRMS_USER_SERACH_CRITERIA_USERTYPE);
+				userSearchCriteria.put(HRMSConstants.HRMS_USER_SEARCH_CRITERA_TENANTID, criteria.getTenantId());
+				userSearchCriteria.put(HRMSConstants.HRMS_USER_SEARCH_CRITERA_USER_SERVICE_UUIDS, criteria.getUserServiceUuids());
+				if(!CollectionUtils.isEmpty(criteria.getNames()))
+					userSearchCriteria.put(HRMSConstants.HRMS_USER_SEARCH_CRITERA_NAME, criteria.getNames().get(0));
+				UserResponse userResponse = userService.getUser(requestInfo, userSearchCriteria);
+				totalCount = userResponse.getTotalCount();
+				userChecked =true;
+				if(!CollectionUtils.isEmpty(userResponse.getUser())) {
+					mapOfUsers.putAll(userResponse.getUser().stream()
+							.collect(Collectors.toMap(User::getUuid, Function.identity())));
+				}
+
+				List<String> uuids = userResponse.getUser().stream().map(User :: getUuid).collect(Collectors.toList());
+				userUUIDs.addAll(uuids);
+
+				if(!CollectionUtils.isEmpty(criteria.getUuids()))
+					criteria.setUuids(criteria.getUuids().stream().filter(userUUIDs::contains).collect(Collectors.toList()));
+				else
+					criteria.setUuids(userUUIDs);
+			}
 		}
-		if(userChecked)
-			criteria.setTenantId(null);
         List <Employee> employees = new ArrayList<>();
 		EmployeeCountResponse employeeCountResponse=new EmployeeCountResponse();
         if(!((!CollectionUtils.isEmpty(criteria.getRoles()) || !CollectionUtils.isEmpty(criteria.getNames()) || !StringUtils.isEmpty(criteria.getPhone())) && CollectionUtils.isEmpty(criteria.getUuids())))
 			employeeCountResponse = repository.fetchEmployees(criteria, requestInfo);
 		employees=employeeCountResponse.getEmployees();
         List<String> uuids = employees.stream().map(Employee :: getUuid).collect(Collectors.toList());
+		if(userChecked)
+			criteria.setTenantId(null);
 		if(!CollectionUtils.isEmpty(uuids)){
             Map<String, Object> userSearchCriteria = new HashMap<>();
 			userSearchCriteria.put(HRMSConstants.HRMS_USER_SERACH_CRITERIA_USERTYPE_CODE, HRMSConstants.HRMS_USER_SERACH_CRITERIA_USERTYPE);
@@ -236,6 +273,14 @@ public class EmployeeService {
 		enrichUser(employee);
 		UserRequest request = UserRequest.builder().requestInfo(requestInfo).user(employee.getUser()).build();
 		try {
+			UserResponse response;
+			if(userService instanceof IndividualService) {
+				String localityCode = (employee.getJurisdictions()!=null && !employee.getJurisdictions().isEmpty())? employee.getJurisdictions().get(0).getBoundary() : null;
+				response = individualService.createUserByLocality(request, localityCode);
+			}
+			else{
+				response = userService.createUser(request);
+			}
 			UserResponse response;
 			if(userService instanceof IndividualService) {
 				String localityCode = (employee.getJurisdictions()!=null && !employee.getJurisdictions().isEmpty())? employee.getJurisdictions().get(0).getBoundary() : null;
@@ -283,8 +328,7 @@ public class EmployeeService {
 	 * @param employee
 	 * @param requestInfo
 	 */
-	private void enrichCreateRequest(Employee employee, RequestInfo requestInfo) {
-
+	private void enrichCreateRequest(String tenantId, Employee employee, RequestInfo requestInfo) {
 		AuditDetails auditDetails = AuditDetails.builder()
 				.createdBy(requestInfo.getUserInfo().getUuid())
 				.createdDate(new Date().getTime())
@@ -298,9 +342,10 @@ public class EmployeeService {
 		});
 		if (employee.getAssignments() != null && !employee.getAssignments().isEmpty()) {
 			employee.getAssignments().stream().forEach(assignment -> {
+//				String tenantId = assignment.getTenantid();
 				assignment.setId(UUID.randomUUID().toString());
 				assignment.setAuditDetails(auditDetails);
-				assignment.setPosition(getPosition());
+				assignment.setPosition(getPosition(tenantId));
 			});
 		}
 		if(!CollectionUtils.isEmpty(employee.getServiceHistory())) {
@@ -341,8 +386,8 @@ public class EmployeeService {
 	 * Fetches next value from the position sequence table
 	 * @return
 	 */
-	public Long getPosition() {
-		return repository.fetchPosition();
+	public Long getPosition(String tenantId) {
+		return repository.fetchPosition(tenantId);
 	}
 
 	/**
@@ -392,7 +437,7 @@ public class EmployeeService {
 		});
 
 		// Push the updated employee request to the HRMS topic for further processing
-		hrmsProducer.push(propertiesManager.getUpdateTopic(), employeeRequest);
+		hrmsProducer.push(tenantId, propertiesManager.getUpdateTopic(), employeeRequest);
 
 		// (Optional) Send reactivation notifications if needed
 		// notificationService.sendReactivationNotification(employeeRequest);